vma reader: drop unused variable
The variable has been unused since commit 67af0fa ("rebased pve
patches") back in 2017. There is no comment to why, but before that,
it was used to error out if there were no disks in the vma archive.
This should be possible however, so it's safe to assume this was an
intentional change.

This fixes compilation with clang19.

Signed-off-by: Fiona Ebner <f.ebner@proxmox.com>