squash some related patches
In particular for backup and savevm-async, a lot can be grouped
together, many patches were just later fixups/improvements. Backup
patches are still grouped in three: original, addition of fleecing
and addition of external backup API (preparatory patches there
could be split and squashed too, but they're relatively new, so that
is something for next time).

The code with all patches applied is still the same.

Signed-off-by: Fiona Ebner <f.ebner@proxmox.com>
Link: https://lore.proxmox.com/20250602102227.120732-3-f.ebner@proxmox.com