bump baseline machine version to 6.0 for Proxmox VE 9
This was missed with QEMU 10.0 and should've been bumped already for
the release of Proxmox VE 9 to get more complete deprecation warnings.

The compile time check for detecting too old baseline is only active
starting with QEMU 10.1, since the new policy on the QEMU side is
enforced starting with that version and that's why it slipped through.

Signed-off-by: Fiona Ebner <f.ebner@proxmox.com>
Link: https://lore.proxmox.com/20250924144848.170589-3-f.ebner@proxmox.com